api for getting category wise votes
