+++ pages/wishlist.js
@@ -1,17 +1,46 @@
 import Head from 'next/head'
 import Link from 'next/link'
 
-export default function Admin() {
+export default function Wishlist() {
+
+  //displaying all the images with category painting from atlas
+  // const [painting, setPainting] = useState([])
+  // const [photography, setPhotography] = useState([])
+  // const [digitalart, setDigitalart] = useState([])
+  // const [theme, setTheme] = useState([])
+  const [isloggedin, setIsLoggedIn] = useState(false)
+
+  const checkifuser = async () => {
+    const loggedInUser = localStorage.getItem("user");
+    //painting images
+    const resh = await fetch(`http://localhost:3000/api/getWishlist/${loggedInUser, "painting"}`)
+    const data = await resh.json()
+    if (data.name !== 'nouser') {
+      setIsLoggedIn(true)
+    }
+    console.log(isloggedin)
+  }
+
+
+
+  // const aggCursor = db.collection("Votes").aggregate(pipeline).toArray()
+  // for await (const doc of aggCursor) {
+  //   console.log(doc);
+  // }
+
+  const getPainting = async () => {
+
+  }
   return (
     <>
       <Head>
-        <title>My Votes</title>
+        <title>My Wishlist</title>
         <meta name="description" content="Generated by create next app" />
         <meta name="viewport" content="width=device-width, initial-scale=1" />
         <link rel="icon" href="/favicon.ico" />
       </Head>
       <div>
-        <h1>This is My Votes Page</h1>
+        <h1>This is My Wishlists Page</h1>
         <Link href="/feedback"><button>Final Submit</button></Link><br></br>
         <div>
           <div>
@@ -28,7 +57,7 @@ export default function Admin() {
           </div>
           <div>
             <p>Theme</p>
-            
+
           </div>
         </div>
       </div>
